Redesign list for booking inquiries
